agent/LLaMaIndex/api/requirements.txt
flask
diskcache
python-dotenv
sqlalchemy
oracledb
llama-index>=0.13.0
llama-index-llms-ollama>=0.6.2
llama-index-embeddings-huggingface>=0.2.0
huggingface-hub>=0.16.4
sentence-transformers>=2.2.2
transformers>=4.35.0
numpy<2.0.0
